+++ packages/stream_core/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Upcoming
+## 0.3.3
 
 ### ✨ Features